% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Plot Population for a Country}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{A character string giving the country name.}
}
\value{
A ggplot object showing population over time.
}
\description{
Generates a population vs. year plot (1950–2020) for a specified country.
}
\examples{
CountryPopulation("Brazil")

}
